Moved `.blender` to `._blender.addon27`
