+++ blenderneuron/_blender/addon/__init__.py
@@ -1,0 +1,201 @@
+# bl_info = {
+#     "name": "BlenderNEURON",
+#     "description": "A Blender GUI for NEURON simulator",
+#     "author": "Justas Birgiolas",
+#     "version": (3, 0),
+#     "blender": (2, 91, 0),
+#     "location": "3D View > Left Side > BlenderNEURON",
+#     "wiki_url": "BlenderNEURON.org",
+#     "tracker_url": "https://github.com/JustasB/BlenderNEURON/issues",
+#     "support": "COMMUNITY",
+#     "category": "Import-Export",
+# }
+
+
+bl_info = {
+    "name": "Add-on Template",
+    "description": "elluuuuuuu",
+    "author": "p2or",
+    "version": (0, 0, 3),
+    "blender": (2, 91, 0),
+    "location": "3D View > Tools",
+    "warning": "", # used for warning icon and text in addons panel
+    "wiki_url": "",
+    "tracker_url": "",
+    "category": "Development"
+}
+
+
+import bpy
+
+from bpy.props import (StringProperty,
+                       BoolProperty,
+                       IntProperty,
+                       FloatProperty,
+                       FloatVectorProperty,
+                       EnumProperty,
+                       PointerProperty,
+                       )
+from bpy.types import (Panel,
+                       Menu,
+                       Operator,
+                       PropertyGroup,
+                       )
+
+
+# ------------------------------------------------------------------------
+#    Scene Properties
+# ------------------------------------------------------------------------
+
+class MyProperties(PropertyGroup):
+
+    my_bool: BoolProperty(
+        name="Enable or Disable",
+        description="A bool property",
+        default = False
+        )
+
+    my_int: IntProperty(
+        name = "Int Value",
+        description="A integer property",
+        default = 23,
+        min = 10,
+        max = 100
+        )
+
+    my_float: FloatProperty(
+        name = "Float Value",
+        description = "A float property",
+        default = 23.7,
+        min = 0.01,
+        max = 30.0
+        )
+
+    my_float_vector: FloatVectorProperty(
+        name = "Float Vector Value",
+        description="Something",
+        default=(0.0, 0.0, 0.0),
+        min= 0.0, # float
+        max = 0.1
+    )
+
+    my_string: StringProperty(
+        name="User Input",
+        description=":",
+        default="",
+        maxlen=1024,
+        )
+
+    my_path: StringProperty(
+        name = "Directory",
+        description="Choose a directory:",
+        default="",
+        maxlen=1024,
+        subtype='DIR_PATH'
+        )
+
+    my_enum: EnumProperty(
+        name="Dropdown:",
+        description="Apply Data to attribute.",
+        items=[ ('OP1', "Option 1", ""),
+                ('OP2', "Option 2", ""),
+                ('OP3', "Option 3", ""),
+               ]
+        )
+
+# ------------------------------------------------------------------------
+#    Operators
+# ------------------------------------------------------------------------
+
+class WM_OT_HelloWorld(Operator):
+    bl_label = "Print Values Operator"
+    bl_idname = "wm.hello_world"
+
+    def execute(self, context):
+        scene = context.scene
+        mytool = scene.my_tool
+
+        # print the values to the console
+        print("Hello World")
+        print("bool state:", mytool.my_bool)
+        print("int value:", mytool.my_int)
+        print("float value:", mytool.my_float)
+        print("string value:", mytool.my_string)
+        print("enum state:", mytool.my_enum)
+
+        return {'FINISHED'}
+
+# ------------------------------------------------------------------------
+#    Menus
+# ------------------------------------------------------------------------
+
+class OBJECT_MT_CustomMenu(bpy.types.Menu):
+    bl_label = "Select"
+    bl_idname = "OBJECT_MT_custom_menu"
+
+    def draw(self, context):
+        layout = self.layout
+
+        # Built-in operators
+        layout.operator("object.select_all", text="Select/Deselect All").action = 'TOGGLE'
+        layout.operator("object.select_all", text="Inverse").action = 'INVERT'
+        layout.operator("object.select_random", text="Random")
+
+# ------------------------------------------------------------------------
+#    Panel in Object Mode
+# ------------------------------------------------------------------------
+
+class OBJECT_PT_CustomPanel(Panel):
+    bl_label = "My Panel"
+    bl_idname = "OBJECT_PT_custom_panel"
+    bl_space_type = "VIEW_3D"
+    bl_region_type = "UI"
+    bl_category = "Tools"
+    bl_context = "objectmode"
+
+
+    @classmethod
+    def poll(self,context):
+        return context.object is not None
+
+    def draw(self, context):
+        layout = self.layout
+        scene = context.scene
+        mytool = scene.my_tool
+
+        layout.prop(mytool, "my_bool")
+        layout.prop(mytool, "my_enum", text="")
+        layout.prop(mytool, "my_int")
+        layout.prop(mytool, "my_float")
+        layout.prop(mytool, "my_float_vector", text="")
+        layout.prop(mytool, "my_string")
+        layout.prop(mytool, "my_path")
+        layout.operator("wm.hello_world")
+        layout.menu(OBJECT_MT_CustomMenu.bl_idname, text="Presets", icon="SCENE")
+        layout.separator()
+
+# ------------------------------------------------------------------------
+#    Registration
+# ------------------------------------------------------------------------
+
+classes = (
+    MyProperties,
+    WM_OT_HelloWorld,
+    OBJECT_MT_CustomMenu,
+    OBJECT_PT_CustomPanel
+)
+
+def register():
+    from bpy.utils import register_class
+    for cls in classes:
+        register_class(cls)
+
+    bpy.types.Scene.my_tool = PointerProperty(type=MyProperties)
+
+def unregister():
+    from bpy.utils import unregister_class
+    for cls in reversed(classes):
+        unregister_class(cls)
+    del bpy.types.Scene.my_tool
+
+__blender__ = dict(bl_info=bl_info, register=register, unregister=unregister)
